Aur helpers (#210)
* Update arch-manjaro.sh

* Update arch-manjaro.sh

+++ setup/arch-manjaro.sh
@@ -13,15 +13,22 @@ sudo pacman -Syyu --noconfirm --needed git git-lfs multilib-devel fontconfig ttf
 # Install android build prerequisites
 echo '[3/4] Installing Android building prerequisites'
 packages="ncurses5-compat-libs lib32-ncurses5-compat-libs aosp-devel xml2 lineageos-devel"
-for package in $packages; do
-    echo "Installing $package"
-    git clone https://aur.archlinux.org/"$package"
-    cd "$package" || exit
-    makepkg -si --skippgpcheck --noconfirm --needed
-    cd - || exit
-    rm -rf "$package"
-done
-
+if command -v paru 2>&1 >/dev/null 
+then
+    paru -S --noconfirm --needed $packages
+elif command -v yay 2>&1 >/dev/null
+then
+    yay -S --noconfirm --needed $packages
+else
+    for package in $packages; do
+        echo "Installing $package"
+        git clone https://aur.archlinux.org/"$package"
+        cd "$package" || exit
+        makepkg -si --skippgpcheck --noconfirm --needed
+        cd - || exit
+        rm -rf "$package"
+    done
+fi
 # Install adb and associated udev rules
 echo '[4/4] Installing adb convenience tools'
 sudo pacman -S --noconfirm --needed android-tools android-udev
